options: remove CONF_TYPE_POSITION
This was the option parser for the off_t C type. These days, off_t is
always int64_t, so replace all its uses by int64_t and CONF_TYPE_INT64.

Fix the --sstep option. It used CONF_TYPE_INT with an off_t variable,
which will result in invalid memory accesses. (It worked, perhaps
because virtually all systems are little endian, and the out of bound
access happened not to overwrite anything important.) Make it use type
double instead, which seems to make more sense for this option.